requirements.txt
# FastAPI and web
fastapi[all]==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
httpx==0.26.0
flask==3.0.2

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis and Celery
redis==5.0.1
celery[redis]==5.3.6

# Settings and logging
pydantic-settings==2.1.0
structlog==24.1.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
scipy==1.12.0
numpy==1.26.4
# praat-parselmouth excluded - has ARM build issues, librosa fallback used

# Cryptography
cryptography==42.0.2

# Authentication
bcrypt==4.1.2
mnemonic==0.21
PyJWT==2.8.0

# Ethereum/Web3
web3==6.15.1
eth-account==0.11.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
